+++ lofty/src/tag/item.rs
@@ -156,7 +156,8 @@ gen_map!(
 	"MUSICBRAINZ_RELEASEGROUPID"     => MusicBrainzReleaseGroupId,
 	"MUSICBRAINZ_ARTISTID"           => MusicBrainzArtistId,
 	"MUSICBRAINZ_ALBUMARTISTID"      => MusicBrainzReleaseArtistId,
-	"MUSICBRAINZ_WORKID"             => MusicBrainzWorkId
+	"MUSICBRAINZ_WORKID"             => MusicBrainzWorkId,
+	"MUSICBRAINZ_ALBUMTYPE"			 => MusicBrainzReleaseType,
 );
 
 gen_map!(
@@ -256,7 +257,8 @@ gen_map!(
 	"MusicBrainz Artist Id"                 => MusicBrainzArtistId,
 	"MusicBrainz Album Artist Id"           => MusicBrainzReleaseArtistId,
 	"MusicBrainz Work Id"                   => MusicBrainzWorkId,
-	"MusicBrainz Album Release Country" 	=> ReleaseCountry
+	"MusicBrainz Album Type"				=> MusicBrainzReleaseType,
+	"MusicBrainz Album Release Country" 	=> ReleaseCountry,
 );
 
 gen_map!(
@@ -343,7 +345,8 @@ gen_map!(
 	"----:com.apple.iTunes:MusicBrainz Artist Id"             => MusicBrainzArtistId,
 	"----:com.apple.iTunes:MusicBrainz Album Artist Id"       => MusicBrainzReleaseArtistId,
 	"----:com.apple.iTunes:MusicBrainz Work Id"               => MusicBrainzWorkId,
-	"----:com.apple.iTunes:MusicBrainz Album Release Country" => ReleaseCountry
+	"----:com.apple.iTunes:MusicBrainz Album Type"		  	  => MusicBrainzReleaseType,
+	"----:com.apple.iTunes:MusicBrainz Album Release Country" => ReleaseCountry,
 );
 
 gen_map!(
@@ -456,7 +459,8 @@ gen_map!(
 	"MUSICBRAINZ_RELEASEGROUPID"              => MusicBrainzReleaseGroupId,
 	"MUSICBRAINZ_ARTISTID"                    => MusicBrainzArtistId,
 	"MUSICBRAINZ_ALBUMARTISTID"               => MusicBrainzReleaseArtistId,
-	"MUSICBRAINZ_WORKID"                      => MusicBrainzWorkId
+	"MUSICBRAINZ_WORKID"                      => MusicBrainzWorkId,
+	"RELEASETYPE"							  => MusicBrainzReleaseType,
 );
 
 macro_rules! gen_item_keys {
@@ -737,6 +741,15 @@ gen_item_keys!(
 		///
 		/// Reference: <https://picard-docs.musicbrainz.org/en/appendices/tag_mapping.html#musicbrainz-work-id>
 		MusicBrainzWorkId,
+
+		/// MusicBrainz Release Type
+		///
+		/// A description of the MusicBrainz release group type (e.g. `album`, `single`, `ep`, etc.).
+		///
+		/// <https://musicbrainz.org/doc/Release_Group/Type>
+		///
+		/// Reference: <https://picard-docs.musicbrainz.org/en/appendices/tag_mapping.html#id32>
+		MusicBrainzReleaseType,
 
 		///////////////////////////////////////////////////////////////